+++ src/main/java/com/movieproject/domain/movie/service/MovieInternalService.java
@@ -1,10 +1,10 @@
-package com.movieproject.movie;
+package com.movieproject.domain.movie.service;
 
-import com.movieproject.common.exception.GlobalException;
 import com.movieproject.domain.director.entity.Director;
+import com.movieproject.domain.director.service.DirectorExternalService;
 import com.movieproject.domain.movie.dto.MovieRequestDto;
+import com.movieproject.domain.movie.dto.MovieResponseDto;
 import com.movieproject.domain.movie.entity.Movie;
-import com.movieproject.domain.movie.exception.MovieErrorCode;
 import com.movieproject.domain.movie.repository.MovieRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -12,13 +12,15 @@ import org.springframework.transaction.annotation.Transactional;
 
 @Service
 @RequiredArgsConstructor
-@Transactional(readOnly = true)
+@Transactional
 public class MovieInternalService {
 
     private final MovieRepository movieRepository;
+    private final DirectorExternalService directorExternalService;
 
-    @Transactional
-    public Movie createMovie(MovieRequestDto.Create requestDto, Director director) {
+    public MovieResponseDto createMovie(MovieRequestDto.Create requestDto) {
+        Director director = directorExternalService.findDirectorById(requestDto.directorId());
+
         Movie movie = Movie.builder()
                 .title(requestDto.title())
                 .releaseDate(requestDto.releaseDate())
@@ -28,11 +30,8 @@ public class MovieInternalService {
                 .director(director)
                 .build();
 
-        return movieRepository.save(movie);
-    }
+        Movie savedMovie = movieRepository.save(movie);
 
-    public Movie findMovieById(Long movieId) {
-        return movieRepository.findById(movieId)
-                .orElseThrow(() -> new GlobalException(MovieErrorCode.MOVIE_NOT_FOUND));
+        return MovieResponseDto.from(savedMovie);
     }
 }
