--- src/test/java/com/movieproject/movie/MovieExternalServiceTest.java
@@ -1,70 +1,0 @@
-package com.movieproject.movie;
-
-import com.movieproject.domain.director.entity.Director;
-import com.movieproject.domain.director.service.DirectorExternalService;
-import com.movieproject.domain.movie.dto.MovieRequestDto;
-import com.movieproject.domain.movie.dto.MovieResponseDto;
-import com.movieproject.domain.movie.entity.Movie;
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.extension.ExtendWith;
-import org.mockito.InjectMocks;
-import org.mockito.Mock;
-import org.mockito.junit.jupiter.MockitoExtension;
-import org.springframework.test.util.ReflectionTestUtils;
-
-import java.time.LocalDate;
-
-import static org.assertj.core.api.Assertions.assertThat;
-import static org.mockito.Mockito.verify;
-import static org.mockito.Mockito.when;
-import static org.mockito.internal.verification.VerificationModeFactory.times;
-
-@ExtendWith(MockitoExtension.class)
-class MovieExternalServiceTest {
-
-    @InjectMocks
-    private MovieExternalService movieExternalService;
-
-    @Mock
-    private MovieInternalService movieInternalService;
-
-    @Mock
-    private DirectorExternalService directorExternalService;
-
-    @Test
-    @DisplayName("영화 등록 성공 테스트")
-    void createMovie_success() {
-        // given
-        Long directorId = 1L;
-        Long movieId = 100L;
-        MovieRequestDto.Create requestDto = new MovieRequestDto.Create(
-                "테스트 영화",
-                LocalDate.of(2025, 1, 1),
-                120,
-                "KOREA",
-                "테스트 장르",
-                1L
-        );
-
-        Director testDirector = Director.of("테스트 감독", "KOREA",LocalDate.of(1969, 9, 14));
-        ReflectionTestUtils.setField(testDirector, "directorId", 1L);
-
-        Movie mockMovie = Movie.builder().director(testDirector).title(requestDto.title()).build();
-        ReflectionTestUtils.setField(mockMovie, "id", movieId);
-
-        when(directorExternalService.findDirectorById(directorId)).thenReturn(testDirector);
-        when(movieInternalService.createMovie(requestDto, testDirector)).thenReturn(mockMovie);
-
-        // when
-        MovieResponseDto responseDto = movieExternalService.createMovie(requestDto);
-
-        // then
-        assertThat(responseDto).isNotNull();
-        assertThat(responseDto.movieId()).isEqualTo(100L);
-        assertThat(responseDto.title()).isEqualTo("테스트 영화");
-        assertThat(responseDto.director().name()).isEqualTo("테스트 감독");
-        verify(directorExternalService, times(1)).findDirectorById(directorId);
-        verify(movieInternalService, times(1)).createMovie(requestDto, testDirector);
-    }
-}
